delete method security done

+++ src/main/java/com/emrekaraman/springsocial/business/concretes/FlowManager.java
@@ -1,10 +1,11 @@
 package com.emrekaraman.springsocial.business.concretes;
 
+import com.emrekaraman.springsocial.auth.userAuthService.UserDetailsManager;
 import com.emrekaraman.springsocial.business.abstracts.FlowService;
-import com.emrekaraman.springsocial.business.abstracts.UserService;
 import com.emrekaraman.springsocial.business.constants.Messages;
 import com.emrekaraman.springsocial.business.dtos.FlowDto;
 import com.emrekaraman.springsocial.business.dtos.PagesDto;
+import com.emrekaraman.springsocial.core.business.BusinessRules;
 import com.emrekaraman.springsocial.core.utilities.*;
 import com.emrekaraman.springsocial.dataAccess.abstracts.FlowDao;
 import com.emrekaraman.springsocial.entities.concretes.Flow;
@@ -17,6 +18,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FlowManager implements FlowService {
@@ -59,7 +61,32 @@ public class FlowManager implements FlowService {
     }
 
     @Override
-    public Result deleteById(Long id) {
+    public DataResult<Flow> findById(Long id) {
+        Optional<Flow> flow = this.flowDao.findById(id);
+        if (!flow.isEmpty()){
+            return new SuccessDataResult(flow.get(),Messages.FLOW_FOUND);
+        }
+        return new ErrorDataResult(Messages.FLOW_NOT_FOUND);
+    }
+
+    @Override
+    public Result isAuthorization(Long flowId,UserDetailsManager userDetailsManager) {
+        Optional<Flow> result = this.flowDao.findById(flowId);
+        if (result.get().getUser().getId() != userDetailsManager.getUser().getId()){
+            return new ErrorResult(Messages.UNAUTHORIZE);
+        }
+        return new SuccessResult();
+    }
+
+    @Override
+    public Result deleteById(Long id, UserDetailsManager userDetailsManager) {
+        Result result = BusinessRules.run(
+                        findById(id),isAuthorization(id,
+                        userDetailsManager
+                ));
+        if (result != null){
+            return new ErrorResult(result.getMessage());
+        }
         this.flowDao.deleteById(id);
         return new SuccessResult(Messages.SUCCESSFULLY_PROCESS);
     }
